fix(poem-openapi-derive): Allow different path param names on same route (#952)
This commits add automatic renaming of path parameters in route table
and param extraction.
When having two same routes with different method and different param
names, it resulted in one of them returning 405 instead of firing the
handler.

This fix is somewhat janky, because it checks if the extracted param
type is `Path` and then it renames is to `param{i}`. There could be
more edge cases down the line, but for now it passess all the tests.

Fixes: https://github.com/poem-web/poem/issues/948
Related: https://github.com/poem-web/poem/issues/429

Signed-off-by: Robin Ilyk <contact@robko.cz>