feat: use enum discriminant name for type name
When generating the different union discriminant, use the name of the
discriminant instead of the name of the type inside the discriminant.
This allows reuse of the type for multiple enumeration discriminants.

BREAKING CHANGE: the type names in the generated OpenAPI spec are
changed, but the contents of the types are the same.

Fixes #1038